polygon to channel index and polygon-based geometry updates
Added functions to index channels to polygons and a new df with polygon ids as index to track changes. Channel geometry in the model is subsequently updated based on changes in polygons (currently hardcoded)

Approach replaces previous function to change the geometry of a complete channel. The sea lever rise function is updated to work with the new approach. The split channel function needs to be updated (this also changes the network gdf as well as the polygon index df, so split channel would also need to update (and reindex?) those).